move CustomizationFieldValidator to Validate ns
