+++ src/Adapter/Product/Validate/CustomizationFieldValidator.php
@@ -26,7 +26,7 @@
 
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Adapter\Product;
+namespace PrestaShop\PrestaShop\Adapter\Product\Validate;
 
 use CustomizationField;
 use PrestaShop\PrestaShop\Adapter\AbstractObjectModelValidator;
