+++ src/PrestaShopBundle/Resources/config/services/adapter/product.yml
@@ -272,13 +272,13 @@ services:
         arguments:
             - '@prestashop.adapter.product.validate.product_validator'
 
-    prestashop.adapter.product.customization_field_validator:
-        class: PrestaShop\PrestaShop\Adapter\Product\CustomizationFieldValidator
+    prestashop.adapter.product.validate.customization_field_validator:
+        class: PrestaShop\PrestaShop\Adapter\Product\Validate\CustomizationFieldValidator
 
     prestashop.adapter.product.customization_field_repository:
         class: PrestaShop\PrestaShop\Adapter\Product\Repository\CustomizationFieldRepository
         arguments:
-            - '@prestashop.adapter.product.customization_field_validator'
+            - '@prestashop.adapter.product.validate.customization_field_validator'
 
     prestashop.adapter.product.command_handler.delete_product_handler:
         class: PrestaShop\PrestaShop\Adapter\Product\CommandHandler\DeleteProductHandler
